requirements.txt
Django==4.2
djangorestframework==3.14.0
numpy==1.24.2
openai==0.27.4
pandas==2.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
sentence_transformers==2.2.2
transformers==4.27.4
